Split character builder derivation hooks
